% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishLoc.R
\docType{data}
\name{FishLocation}
\alias{FishLocation}
\title{Fish Location Data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 15 rows and 6 columns.}
\source{
Data Collected From : Many lovely ports in Ireland

Methods Used : Pursued the Irish fishermen with mugs of beers to get the fish count
}
\usage{
FishLocation
}
\description{
Data collected as part of Fisheries Management Course , ESM 707 Bren School.
}
\details{
Fish Of Ireland

 Purpose of Data Collection -> To test Charles Haas's saying ,

"Give a man a fish, and he can eat for a day.

 But teach a man how to fish, and he'll be dead of mercury poisoning inside of three years."
}
\author{
Anu Balakrishnan
}
\keyword{datasets}
