% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishPortPrice.R
\name{FishPortPrice}
\alias{FishPortPrice}
\title{Function that takes as two inputs
1) a table that has prices for different fish}
\usage{
FishPortPrice(df1, df2)
}
\arguments{
\item{df1}{the table with fish name and proce
2) a table that has the number caught for each fish species for each location}

\item{df2}{The table with fish name and fish caught per location
Data table requriements :  each location is in a different column and each fish is in a different row
Function Output}
}
\value{
Port Names and Highest number of fish caught per location

rowPriceLocTot : total revenue for each location

rowPriceLocTotalAll : total fisheries revenue sum

if user requests it graph of revenue by location and total revenue (as text)
}
\description{
Function that takes as two inputs
1) a table that has prices for different fish
}
