% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishTypeTot.R
\name{FishTypeTot}
\alias{FishTypeTot}
\title{Function that takes fish data}
\usage{
FishTypeTot(df1)
}
\arguments{
\item{df1}{fish data with fish name in atleast one column}
}
\value{
numRowdf1 total number of unique fish in the dataset

End of comment
}
\description{
Function that takes fish data
}
